Refactor Agent to use prototype strategy
